feat: update global styles
This commit updates the global styles in the _global.scss file. The changes include:

Fixed a typo in the class selector for the body element, changing .body to body.
Updated the fill color of the .logo SVG element to #ff39bc on hover.
Updated the fill color of the .title element to #ff39bc on hover.
Added new styles for the .button class, including custom arrow icons and hover effects.
Added a new class .button--back to style the back button in the booking form.
Removed the console.log statement from the previously deleted faq.js component.
These changes enhance the visual appearance and interactivity of the application, making it more user-friendly and appealing.