style: improve designs section styles
In this commit, the styles for the designs section are improved in the _designs.scss file. The changes include the following:

Removed unnecessary spaces and empty lines for better code readability.

Added a transition effect to the images inside the designs section. When hovering over the images, they will now smoothly scale up by 10% using the transform property.

Adjusted the size of the first image in the designs section to have a width of 200px and a height of 199px. These dimensions ensure a better display and consistency with the other designs.